Relax the signature check in GetBrushSectionOffsets
It now breaks out of the loop if the signature check fails instead of
returning null.

The reason for this change is that some files have a descriptor section
length that is too short, it points to an area within the descriptor
data instead of the end of the file.