Added a palette combobox that loads the current Paint.NET palette by default, and loads additional palettes from the default/custom directories the user specifies. The GUI that lets the user set brush image directories now includes palette directories below. The combobox also includes generated palettes following well-known color schemes, which live-update as the user switches color. Even uses HSLuv color space for the primary-to-secondary gradient.
Extended color logic in shortcut syntax to allow arithmetic in ARGB or AHSV. Fixed various discrepancies and issues in some of the new components / UI. Included was a fix to add-wrap and sub-wrap that users might've experienced from earlier. Code refactored a bit to generate pressure control GUI instead of exhaustively writing it out. (Shortcut targets for pressure need to be reworked because number of constraint types will be >1 in the future. Tab pressure is just one constraint type.)