Add InteropBitmap so that IBitmap can be used as S.D.Bitmap where needed
`bmpCommitted` is mostly used in methods that operate on `S.D.Bitmap`. However, the new effect system in PDN 5.0+ operates on `IBitmap`s. In order to safely interoperate between these, without having expensive copies, the bitmap must be allocated as an `IBitmap` with an aliased `S.D.Bitmap`.

This ensures that lifetime can be controlled by the `IBitmap`, which supports reference counting (via `CreateRef()`/`TryCreateRef()`). `S.D.Bitmap` has no way to link itself to another object, so the constructor overload that takes a pointer to bitmap data is very prone to use-after-free errors. The `IBitmap` implements a "link" from the `S.D.Bitmap` to itself via `ConditionalWeakTable`, which ensures that the `IBitmap` won't be garbage collected if the only reference to it is through the "link" from `S.D.Bitmap`. (so you could create the `InteropBitmap`, then use `AsGdipBitmap()`, and then forget/drop the reference to the `InteropBitmap`, and it still wouldn't get freed until the `S.D.Bitmap` is garbage collected.)

Also, PDN does not provide a built-in way to create a "shared" `IBitmap` that points to memory allocated elsewhere (for good reason -- it immediately causes major bug hazards that I did see tripped over during PDN 5.0's private alpha period!). So this custom mechanism was required.